+++ dp_tools/__init__.py
@@ -26,4 +26,4 @@ Framework overview:
 See README for installation and CLI usage.
 """
 
-__version__ = "1.3.11"
+__version__ = "1.3.12"