+++ dp_tools/core/configuration.py
@@ -33,7 +33,7 @@ def load_full_config(config: Union[str, Path]) -> dict:
 
 def load_config(config: Union[tuple[str, str], Path]) -> dict:
     """Load yaml configuration file. Allows loading from either:
-      - A prepackaged configuration file using a tuple of ('config_type','config_version') (e.g. ('bulkRNASeq','Latest'), ('microarray','0'))
+      - A prepackaged configuration file using a tuple of ('config_type','config_version') (e.g. ('bulkRNASeq','Latest'), ('microarray_agilent','Latest'))
       - A configuration file supplied as a Path object
 
     :param config: Configuration file to load